Fixed get/set-listener-addrport test vectors

+++ model/functions/set-IPv4.go
@@ -43,7 +43,7 @@ var SetIPv4 = Function{
 						Name: "controller",
 						Type: "uint32",
 					},
-					Value: uint32(405419896),
+					Value: 405419896,
 				},
 				{
 					Arg: Arg{
