+++ model/functions/set-door-passcodes.go
@@ -49,7 +49,7 @@ var SetDoorPasscodes = Function{
 						Name: "controller",
 						Type: "uint32",
 					},
-					Value: uint32(405419896),
+					Value: 405419896,
 				},
 				{
 					Arg: Arg{
