+++ model/functions/set-event-index.go
@@ -31,7 +31,7 @@ var SetEventIndex = Function{
 						Name: "controller",
 						Type: "uint32",
 					},
-					Value: uint32(405419896),
+					Value: 405419896,
 				},
 				{
 					Arg: Arg{
