+++ model/functions/set-listener.go
@@ -43,7 +43,7 @@ var SetListener = Function{
 						Name: "controller",
 						Type: "uint32",
 					},
-					Value: uint32(405419896),
+					Value: 405419896,
 				},
 				{
 					Arg: Arg{
